feat!: implement pino@7 transport interface
Pino has moved toward in-process transports via worker threads in the
latest version. This refactors the stream implementation to work with
this new interface. The "legacy" cli option is still available for users
that would like to continue sending logs from a separate process.

BREAKING CHANGE: This transport is now untested *as an in-process log
destination* for `pino < 7.0.0`, and while it may work, it is not recommended
and will no longer be officially supported. Users that wish to use the
latest version of `pino-logdna` with earlier versions of `pino` should
pipe program output to the transport instead. See
https://getpino.io/#/docs/transports?id=legacy-transports

Closes #3